packages/pivot/pyproject.toml
[project]
name = "pivot"
version = "0.1.0"
description = "High-performance Python pipeline tool with automatic code change detection"
requires-python = ">=3.13,<3.14"
license = { text = "TBD" }
authors = [{ name = "Internal Team" }]
dependencies = [
  "anyio>=4.0",
  "click>=8.0",
  "deepmerge>=2.0",
  "dulwich>=0.22.0",
  "filelock>=3.12.0",
  "flatten-dict>=0.4.0",
  "grandalf>=0.8",
  "lmdb>=1.4.0",
  "loky>=3.4",
  "networkx>=3.0",
  "pathspec>=0.12.0",
  "pandas>=2.0",
  "pydantic>=2.0",
  "pygtrie>=2.5.0",
  "pyyaml>=6.0",
  "rich>=13.0",
  "ruamel.yaml>=0.18.0",
  "tabulate>=0.9.0",
  "tqdm>=4.66",
  "watchfiles>=1.0",
  "xxhash>=3.0",
]

[project.optional-dependencies]
dvc = ["dvc>=3.0"]
matplotlib = ["matplotlib>=3.5"]
s3 = ["aioboto3>=13.0.0"]
tui = ["pivot-tui"]
docs = [
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.5.0",
  "mkdocstrings[python]>=0.26.0",
  "mkdocs-gen-files>=0.5.0",
  "mkdocs-literate-nav>=0.6.0",
  "mkdocs-section-index>=0.3.9",
  "griffe>=1.3.0",
]
dev = [
  "basedpyright>=1.0",
  "botocore-stubs>=1.35.0",
  "joblib>=1.4.0",
  "joblib-stubs>=1.5.0",
  "matplotlib-stubs>=0.3",
  "mypy>=1.19.1",
  "pandas-stubs>=2.0",
  "pivot[dvc]",
  "pivot[matplotlib]",
  "pivot[s3]",
  "pytest-aioboto3>=0.1.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=4.0",
  "pytest-mock>=3.15.1",
  "pytest-rerunfailures>=16.1",
  "pytest-watcher>=0.6.2",
  "pytest-xdist>=3.8.0",
  "pytest>=7.0",
  "ruff>=0.8.0",
  "types-aioboto3[all]>=13.0.0",
]

[project.scripts]
pivot = "pivot.cli:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
pivot-tui = { workspace = true }
